requirements.txt
torch>=2.2
torchvision>=0.17
tensorboard>=2.14
numpy>=1.24
pandas>=2.0
Pillow>=10.0
PyYAML>=6.0
hydra-core>=1.3
omegaconf>=2.3
scikit-learn>=1.3
matplotlib>=3.8
opencv-python>=4.9.0
scikit-image>=0.22.0
tqdm>=4.66
ipykernel
